requirements.txt
torch==1.9.0
pytorch-lightning==1.4.4
torchio==0.18.72
monai==0.9.0
py-topping==0.4.0
Office365-REST-Python-Client==2.3.13
PyYAML==6.0
pylint==2.15.0
pandas==1.4.3
